feat(shared/rules): extract JSON-DSL to shared/rules/dsl
Move the rule interpreter out of svc-04 internal so svc-08 can reuse it
without violating Go internal import boundaries. SVC-04 keeps the same
exported names via aliases to dsl.Evaluate and dsl types.
