+++ services/svc-04-header-analysis/internal/rules/types.go
@@ -1,26 +1,18 @@
 // Package rules holds the rule cache, DSL interpreter, and evaluator for
-// SVC-04 Header Analysis. The same primitives can be reused by SVC-08.
+// SVC-04 Header Analysis. The DSL itself now lives in
+// shared/rules/dsl so it can be reused by SVC-08 and future analysers;
+// this package keeps the SVC-04-specific concerns (signal snapshot
+// from header signals, category-aware Evaluator, three-tier cache).
 package rules
 
-import "encoding/json"
+import (
+	"github.com/saif/cybersiren/shared/rules/dsl"
+)
 
 // CachedRule is the in-memory shape of a rule loaded from Postgres.
-//
-// Each rule definition is immutable once loaded — the cache replaces
-// the entire rule set on refresh, so callers can hold a slice of these
-// safely without locking until a refresh swaps it.
-type CachedRule struct {
-	ID          int64           `json:"id"`
-	OrgID       *int64          `json:"org_id,omitempty"` // nil = global rule
-	Name        string          `json:"name"`
-	Version     string          `json:"version"`
-	Target      string          `json:"target"`
-	ScoreImpact int             `json:"score_impact"`
-	Logic       json.RawMessage `json:"logic"`
-}
+// Aliased to the shared DSL type so SVC-08 and SVC-04 read the same
+// rule shape without conversion.
+type CachedRule = dsl.CachedRule
 
 // FiredRule is the output of evaluating a single rule that matched.
-type FiredRule struct {
-	Rule        CachedRule
-	MatchDetail json.RawMessage
-}
+type FiredRule = dsl.FiredRule
